move git content from internal in modules to internal in agent
