+++ internal/vcs/auth.go
